Remove dependency on juju/loggo in favor of stdlib
This commit removes the dependency on GPLv3 package `juju/loggo` and
instead prefers to depend on the standard library logger. In order to
retain the functionality of (sort of) levelled logging, the
`http.Client` struct now has a `SetTrace(bool)` method which will enable
or disable printing of request traces. This is not thread safe, but if
that is an issue it would be reasonably straightforward to synchronize
access.